OZ-351: Migrate from net.mekomsolutions → com.ozonehis namespace.
