+++ src/test/java/com/ozonehis/eip/routes/senaite/AuthenticateToSenaiteRouteITest.java
@@ -1,4 +1,4 @@
-package net.mekomsolutions.eip.routes.senaite;
+package com.ozonehis.eip.routes.senaite;
 
 import static org.junit.Assert.assertEquals;
 
@@ -13,11 +13,11 @@ import org.springframework.context.annotation.Import;
 
 @MockEndpoints
 @Import({ TestConfiguration.class})
-public class AuthenticateToOpenmrsRouteITest extends BaseWatcherRouteTest {
-    
-    @Before
+public class AuthenticateToSenaiteRouteITest extends BaseWatcherRouteTest {  
+
+	@Before
     public void setup() throws Exception {
-    	loadXmlRoutesInDirectory("senaite", "authenticate-to-openmrs-route.xml");
+    	loadXmlRoutesInDirectory("senaite", "authenticate-to-senaite-route.xml");
     }
 
 	@Test
@@ -34,10 +34,11 @@ public class AuthenticateToOpenmrsRouteITest extends BaseWatcherRouteTest {
     	exchange.getIn().setBody(event);
     	
     	// replay
-    	producerTemplate.send("direct:authenticate-to-openmrs", exchange);
+    	producerTemplate.send("direct:authenticate-to-senaite", exchange);
     	
     	// verify
-    	assertEquals("Basic c3VwZXJtYW46QWRtaW4xMjM=", exchange.getIn().getHeader("Authorization"));
+    	assertEquals("Basic YWRtaW46YWRtaW4=", exchange.getIn().getHeader("Authorization"));
     	assertEquals(event, exchange.getIn().getBody());
     }
+
 }
